requirements.txt
alembic==1.9.0
Babel==2.11.0
blinker==1.5
click==8.1.3
dependency-injector==4.40.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
gunicorn==20.1.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pytz==2022.6
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
